Lockheed1049h: re-model the engines completely for better historical accuracy
The modeling of the engines was a bit inaccurate, partly because of
inconsistent data sources and partly due to old limitations of JSBsim.
This new model uses consistent historical data, assuming 135/145 octane
fuel which is no longer available anywhere.  The data used to tune the
model are included with references.

git-svn-id: http://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/Lockheed1049h@351 9eb9bc84-db9d-4e05-893b-26dd41ac94a6