Lockheed1049h: Remove artificial boosting of attitude indicator gyro
Removed timer loop that artificially boosted AI spin because the AI
appears to be simulated properly without it. This code may have been
working around an old problem that has gone away.

To maintain accuracy in the AI, the spin must be above 0.8. To maintain
spin above 0.8, the vacuum system must provide a suction above 4.0. At
low RPM it doesn't do so, but once the engines are above idle, the AI
works correctly and decays slowly when engines return to idle.

Refer to $FLIGHTGEAR/src/Instrumentation/attitude-indicator.cxx and
$FLIGHTGEAR/src/Instrumentation/gyro.cxx.

Patch by Richard Senior.

git-svn-id: https://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/Lockheed1049h@593 9eb9bc84-db9d-4e05-893b-26dd41ac94a6