Lockheed1049h version 2.7: Shortcut keys and dialog set autopilot modes independently
The Lockheed1049h autopilot, as controlled by the cockpit model, does not
have independently selected lateral and vertical modes. When the autopilot
is on, both the vertical and horizontal axes are controlled by the autopilot.

When using shortcut keys and the generic autopilot dialog, Flightgear pilots
are used to being able to set modes independently, e.g. Ctrl-H to switch on
heading hold only. This change allows that to happen at the expense of a
consistent and realistic state in the cockpit autopilot.

git-svn-id: http://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/Lockheed1049h@635 9eb9bc84-db9d-4e05-893b-26dd41ac94a6