Lockheed1049h version 2.10: Introduce refuelling errors
On the real aircraft, tanks were refuelled individually and measured using a
combination of internal and external dipsticks, so loading exact quantities and
having an exact balance between tanks is not realistic.

This commit introduces a configurable error percentage when refuelling, which
provides more realistic refuelling tolerances. It also leads to more realistic
fuel management during a flight because opposite tanks need to be managed into a
state of balance.

git-svn-id: https://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/Lockheed1049h@666 9eb9bc84-db9d-4e05-893b-26dd41ac94a6