Lockheed1049h version 3.0: rewrite multiplayer support entirely for a pilot and a flight engineer.
Instead of a pilot and a copilot who cannot control the aircraft, we
now have a pilot and a flight engineer who can control throttle,
mixture, superchargers, cowl flaps, magnoetos, engine start, fuel
valves including crossfeed and jettison.

The multiplayer protocol in this version is incompatible with previous
versions of the Connie, hence the major version number.

git-svn-id: http://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/Lockheed1049h@2709 9eb9bc84-db9d-4e05-893b-26dd41ac94a6